fix x,y type of MoveData
MoveDataにあるx座標とy座標の方をfloatからintに変更した。そのほか影響のある範囲も修正していった。

+++ Assets/Scripts/MoveData.cs
@@ -8,7 +8,7 @@ public class MoveData
     public int pieceType;     // 駒の種類
     public int rotation;      // 回転
     public bool flipped;      // 反転
-    public float x;           // X座標
-    public float y;           // Y座標
+    public int x;           // X座標
+    public int y;           // Y座標
     public bool touchdown;    // タッチダウン
 }
